+++ CHANGES.md
@@ -13,6 +13,7 @@
 * Runtime: fix path normalization (#1848)
 * Runtime: fix reading from the pseudo-filesystem (#1859)
 * Runtime: fix initialization of standard streams under Windows (#1849)
+* Compiler: fix stack overflow issues with double translation (#1869)
 
 # 6.0.1 (2025-02-07) - Lille